+++ aim/web/ui_v2/src/pages/Home/Home.scss
@@ -1,4 +1,4 @@
-.Front__container {
+.Home__container {
   background-color: #FFFFFF;
   overflow: hidden;
   h2 {
@@ -19,11 +19,11 @@
   margin: 0;
 }
 
-.Front__Explore__container {
+.Home__Explore__container {
   border-top: 1px solid #E8F1FC;
   display: flex;
 }
 
-.Front__Activity__container {
+.Home__Activity__container {
   padding: 1.5em;
 }
