+++ aim/web/ui_v2/src/pages/Home/Home.tsx
@@ -3,23 +3,23 @@ import React from 'react';
 import ExploreAim from './components/ExploreAim/ExploreAim';
 import SetupGuide from './components/SetupGuide/SetupGuide';
 import Activity from './components/Activity/Activity';
-import { IFrontProps } from 'types/pages/front/Front';
+import { IFrontProps } from 'types/pages/home/Home';
 
-import './frontStyle.scss';
+import './Home.scss';
 
-function Front({
+function Home({
   activityData,
 }: IFrontProps): React.FunctionComponentElement<React.ReactNode> {
   return (
-    <section className='Front__container'>
-      <div className='Front__Activity__container'>
+    <section className='Home__container'>
+      <div className='Home__Activity__container'>
         <Activity activityData={activityData} />
       </div>
-      <div className='Front__Explore__container'>
+      <div className='Home__Explore__container'>
         <SetupGuide />
         <ExploreAim />
       </div>
     </section>
   );
 }
-export default Front;
+export default Home;
