+++ aim/web/ui_v2/src/pages/Home/components/Activity/Activity.scss
