+++ aim/web/ui_v2/src/pages/Home/components/CodeBlock/CodeBlock.scss
