+++ aim/web/ui_v2/src/pages/Home/components/ExploreAimCard/ExploreAimCard.tsx
@@ -3,7 +3,7 @@ import { NavLink } from 'react-router-dom';
 
 import { IExploreCard } from '../ExploreAim/ExploreAim';
 
-import './exploreAimCardStyle.scss';
+import './ExploreAimCard.scss';
 
 function ExploreAimCard({
   title,
