+++ aim/web/ui_v2/src/pages/Home/components/SetupGuide/SetupGuide.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 
 import CodeBlock from '../CodeBlock/CodeBlock';
 
-import './setupGuideStyle.scss';
+import './SetupGuide.scss';
 
 function SetupGuide(): React.FunctionComponentElement<React.ReactNode> {
   return (
