+++ aim/web/ui_v2/src/routes/routes.tsx
@@ -6,13 +6,13 @@ const Params = React.lazy(() => import('pages/Params/ParamsContainer'));
 const Bookmarks = React.lazy(
   () => import('pages/Bookmarks/BookmarksContainer'),
 );
-const Front = React.lazy(() => import('pages/Front/FrontContainer'));
+const Home = React.lazy(() => import('pages/Home/HomeContainer'));
 const TagsContainer = React.lazy(() => import('pages/Tags/TagsContainer'));
 const CreateTag = React.lazy(() => import('pages/Tags/CreateTag'));
 const TagDetail = React.lazy(() => import('pages/Tags/TagDetail'));
 
 const PATHS = {
-  FRONT: '/front',
+  HOME: '/',
   RUNS: '/runs',
   METRICS: '/metrics',
   METRICS_ID: '/metrics/:appId',
@@ -25,8 +25,8 @@ const PATHS = {
 
 const routes = {
   FRONT: {
-    path: PATHS.FRONT,
-    component: Front,
+    path: PATHS.HOME,
+    component: Home,
     showInSidebar: false,
     displayName: null,
   },
