+++ aim/web/ui_v2/src/types/pages/home/Home.d.ts
