+++ aim/web/ui_v2/src/types/pages/home/components/Activity/Activity.d.ts
@@ -1,4 +1,4 @@
-import { IActivityData } from 'types/pages/front/components/Front';
+import { IActivityData } from 'types/pages/home/components/Home';
 
 export interface IActivityProps {
   activityData: IActivityData;
